backend/pyproject.toml
[project]
name = "iivf-chatbot-backend"
version = "0.1.0"
description = "AI-powered backend for IVF chatbot"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.20.0",
    "beanie>=1.29.0",
    "pydantic[email]>=2.11.3",
    "pydantic-settings>=2.8.1",
    "bcrypt==4.0.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "boto3>=1.40.18",
    "geopy>=2.4.1",
    "haversine>=2.9.0",
    "motor>=3.3.1",
    "loguru>=0.7.2",
    "redis>=5.0.0", 
    "openai>=1.107.1",
    "langchain>=0.3.27",
    "langchain-community==0.3.29",
    "langchain-core==0.3.76",
    "langchain-text-splitters==0.3.11",
    "langsmith==0.4.30",
    "lazy-model==0.3.0",
    "loguru==0.7.3",
    "lxml==6.0.2",
    "chromadb ==1.1.0",
    "pypdf ==3.14.0",
    "langchain-openai>=0.3.33",
    "jq==1.10.0",
    "rank-bm25==0.2.2",
    "aiosmtplib==5.0.0",
    "azure-identity>=1.25.1",
    "azure-keyvault-secrets>=4.10.0"
]
